# api/openapi/mcp-gateway.yaml
openapi: 3.0.3
info:
  title: MCPEG - Model Context Protocol Enablement Gateway API
  description: |
    API-first gateway providing MCP protocol access to various backend services.
    
    This API follows the Model Context Protocol (MCP) specification while providing
    enhanced features for service discovery, routing, and management.
    
    **Key Features:**
    - Full MCP 2025-03-26 protocol compliance
    - Multi-service adapter support
    - Configuration-driven service routing
    - Enhanced security and validation
    - LLM-optimized error reporting
    - Comprehensive health monitoring
    
    **API-First Design:**
    All code is generated from this OpenAPI specification following the
    single source of truth principle.
  version: "1.0.0"
  contact:
    name: MCPEG API Support
    url: https://github.com/yourusername/mcpeg
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.mcpeg.local
    description: Production server
  - url: https://staging-api.mcpeg.local
    description: Staging server
  - url: http://localhost:8080
    description: Development server

# Security schemes for the API
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - MCPAuth: []

# Global security requirements
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: X-API-Key
    description: API key for service authentication
  BearerAuth:
    type: http
    scheme: bearer
    bearerFormat: JWT
    description: JWT token for user authentication
  MCPAuth:
    type: http
    scheme: bearer
    bearerFormat: MCP
    description: MCP-specific authentication token

# Common response headers
...
paths:
...
  /api/services:
    get:
      operationId: listServices
      summary: List configured services
      description: |
        List all configured service adapters with their status and capabilities.
        This is a management API for service discovery and monitoring.
      tags:
        - Service Management
      parameters:
        - name: status
          in: query
          description: Filter services by status
          schema:
            type: string
            enum: [active, inactive, error]
        - name: type
          in: query
          description: Filter services by adapter type
          schema:
            type: string
      responses:
        '200':
          description: List of configured services
          headers:
            <<: *common-response-headers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: registerService
      summary: Register a new service
      description: |
        Register a new service adapter with the gateway.
        This allows dynamic service registration at runtime.
      tags:
        - Service Management
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceRegistrationRequest'
      responses:
        '201':
          description: Service registered successfully
          headers:
            <<: *common-response-headers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRegistrationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/ServiceConflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/services/{serviceId}:
    get:
      operationId: getService
      summary: Get service details
      description: Get detailed information about a specific service
      tags:
        - Service Management
      parameters:
        - name: serviceId
          in: path
          required: true
          description: Unique service identifier
          schema:
            type: string
          example: "mysql-prod"
      responses:
        '200':
          description: Service details
          headers:
            <<: *common-response-headers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetails'
        '404':
          $ref: '#/components/responses/ServiceNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      operationId: updateService
      summary: Update service configuration
      description: Update the configuration of an existing service
      tags:
        - Service Management
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceUpdateRequest'
      responses:
        '200':
          description: Service updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceDetails'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/ServiceNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      operationId: unregisterService
      summary: Unregister a service
      description: Remove a service from the gateway
      tags:
        - Service Management
      parameters:
        - name: serviceId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Service unregistered successfully
          headers:
            <<: *common-response-headers
        '404':
          $ref: '#/components/responses/ServiceNotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # Health and Monitoring API
...
components:
  schemas:
...
    JSONRPCResponse:
      type: object
      required:
        - jsonrpc
        - id
      properties:
        jsonrpc:
          type: string
          enum: ["2.0"]
        id:
          oneOf:
            - type: string
            - type: number
            - type: "null"
        result:
          type: object
          description: Result of successful method call
        error:
          $ref: '#/components/schemas/JSONRPCError'
    JSONRPCError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Error code
        message:
          type: string
          description: Error message
        data:
          type: object
          description: Additional error data

    # MCP Protocol schemas
...
    Tool:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Tool name
          example: "mysql_query"
        description:
          type: string
          description: Tool description
          example: "Execute SQL queries against MySQL database"
        inputSchema:
          type: object
          description: JSON schema for tool input parameters
          example:
            type: object
            properties:
              query:
                type: string
                description: SQL query to execute
              database:
                type: string
                description: Target database name
            required: ["query"]
...
    Resource:
      type: object
      required:
        - uri
        - name
      properties:
        uri:
          type: string
          format: uri
          description: Resource URI
          example: "file:///home/user/config.yaml"
        name:
          type: string
          description: Human-readable resource name
          example: "Configuration File"
        description:
          type: string
          description: Resource description
        mimeType:
          type: string
          description: Resource MIME type
          example: "application/yaml"
...
    Prompt:
      type: object
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: Prompt name
          example: "sql_query_helper"
        description:
          type: string
          description: Prompt description
        arguments:
          type: array
          items:
            type: object
            required:
              - name
              - description
            properties:
              name:
                type: string
              description:
                type: string
              required:
                type: boolean
                default: false
...
    ServiceAdapter:
      type: object
      required:
        - id
        - name
        - type
        - status
      properties:
        id:
          type: string
          description: Unique service identifier
          example: "mysql-prod"
        name:
          type: string
          description: Human-readable service name
          example: "Production MySQL Database"
        type:
          type: string
          description: Service adapter type
          example: "mysql"
          enum: ["mysql", "postgresql", "weather", "file", "git", "memory"]
        status:
          type: string
          enum: ["active", "inactive", "error"]
        endpoint:
          type: string
          description: Service endpoint URL
        version:
          type: string
          description: Service version
        capabilities:
          type: object
          description: Service capabilities
        configuration:
          type: object
          description: Service configuration (sanitized)
        lastHealthCheck:
          type: string
          format: date-time
        metadata:
          type: object
          description: Additional service metadata
    ServiceListResponse:
      type: object
      required:
        - services
        - total
      properties:
        services:
          type: array
          items:
            $ref: '#/components/schemas/ServiceAdapter'
        total:
          type: integer
          description: Total number of services
        page:
          type: integer
          description: Current page number
        pageSize:
          type: integer
          description: Number of items per page
    ServiceRegistrationRequest:
      type: object
      required:
        - name
        - type
        - configuration
      properties:
        name:
          type: string
          description: Service name
        type:
          type: string
          description: Service adapter type
        configuration:
          type: object
          description: Service configuration
        metadata:
          type: object
          description: Additional metadata
    ServiceRegistrationResponse:
      type: object
      required:
        - id
        - status
      properties:
        id:
          type: string
          description: Generated service ID
        status:
          type: string
          enum: ["registered", "pending", "error"]
        message:
          type: string
          description: Registration status message
    ServiceDetails:
      allOf:
        - $ref: '#/components/schemas/ServiceAdapter'
        - type: object
          properties:
            tools:
              type: array
              items:
                $ref: '#/components/schemas/Tool'
            resources:
              type: array
              items:
                $ref: '#/components/schemas/Resource'
            prompts:
              type: array
              items:
                $ref: '#/components/schemas/Prompt'
            healthStatus:
              $ref: '#/components/schemas/HealthStatus'
    ServiceUpdateRequest:
      type: object
      properties:
        name:
          type: string
        configuration:
          type: object
        metadata:
          type: object

    # Health and Monitoring schemas
    HealthCheck:
      type: object
      required:
        - name
        - status
        - timestamp
      properties:
        name:
          type: string
          description: Health check name
        status:
          type: string
          enum: ["healthy", "degraded", "unhealthy", "unknown"]
        message:
          type: string
          description: Status message
        duration:
          type: string
          description: Check duration
        details:
          type: object
          description: Additional check details
        suggestions:
          type: array
          items:
            type: string
          description: Suggested actions for issues
    HealthStatus:
      type: object
      required:
        - status
        - timestamp
        - version
        - uptime
      properties:
        status:
          type: string
          enum: ["healthy", "degraded", "unhealthy", "unknown"]
        timestamp:
          type: string
          format: date-time
        version:
          type: string
        uptime:
          type: string
          description: System uptime duration
        checks:
          type: array
          items:
            $ref: '#/components/schemas/HealthCheck'
        summary:
          type: object
          properties:
            total:
              type: integer
            healthy:
              type: integer
            degraded:
              type: integer
            unhealthy:
              type: integer
            critical:
              type: integer
        suggestions:
          type: array
          items:
            type: string
...
    Error:
      type: object
      required:
        - code
        - message
        - timestamp
      properties:
        code:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          description: Additional error details
        suggestions:
          type: array
          items:
            type: string
          description: Suggested remediation actions
        traceId:
          type: string
          description: Request trace identifier
        requestId:
          type: string
          description: Request identifier
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/JSONRPCResponse'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/JSONRPCError'
                      - type: object
                        properties:
                          code:
                            type: integer
                            enum: [-32602]
                          message:
                            type: string
                            example: "Invalid params"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Authentication required"
            timestamp: "2025-07-11T10:30:00Z"
            suggestions:
              - "Provide valid authentication credentials"
              - "Check API key or token validity"
...
    ServiceNotFound:
      description: Service not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ServiceConflict:
      description: Service already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/JSONRPCResponse'
              - type: object
                properties:
                  error:
                    allOf:
                      - $ref: '#/components/schemas/JSONRPCError'
                      - type: object
                        properties:
                          code:
                            type: integer
                            enum: [-32603]
                          message:
                            type: string
                            example: "Internal error"

# API documentation tags
...